backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
numpy>=2.0.0
scikit-learn>=1.5.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
httpx>=0.27.0
python-multipart>=0.0.12
geopy>=2.4.0

